синхронизация - работа с сервером вынесена в отдельный модуль, подготовлен каркас модуля для конвертирования данных, отправка текущего конфига

+++ src/cfg/webjoin.h
@@ -30,7 +30,7 @@ class ConfigWebJoin : public Config<cfg_webjoin_t> {
         ConfigWebJoin();
         ConfigWebJoin(uint32_t authid, uint32_t secnum);
         
-        uint8_t authid() const { return data.authid; }
+        uint32_t authid() const { return data.authid; }
 };
